+++ tmplORM.types.hxx
@@ -1,7 +1,6 @@
 #ifndef tmplORM_TYPES__HXX
 #define tmplORM_TYPES__HXX
 
-#include <cinttypes>
 #include <conversions.hxx>
 
 /*!
@@ -79,7 +78,7 @@ namespace tmplORM
 				uint16_t _hour;
 				uint16_t _minute;
 				uint16_t _second;
-				uint64_t _nanoSecond;
+				uint32_t _nanoSecond;
 
 				/*!
 				* @brief Raise 10 to the power of power.
@@ -93,7 +92,7 @@ namespace tmplORM
 			public:
 				constexpr ormTime_t() noexcept : _hour(0), _minute(0), _second(0), _nanoSecond(0) { }
 				constexpr ormTime_t(const uint16_t hour, const uint16_t minute, const uint16_t second,
-					const uint64_t nanoSecond) noexcept : _hour{hour}, _minute{minute},
+					const uint32_t nanoSecond) noexcept : _hour{hour}, _minute{minute},
 					_second{second}, _nanoSecond{nanoSecond} { }
 
 				constexpr uint16_t hour() const noexcept { return _hour; }
@@ -102,8 +101,8 @@ namespace tmplORM
 				void minute(const uint16_t minute) noexcept { _minute = minute; }
 				constexpr uint16_t second() const noexcept { return _second; }
 				void second(const uint16_t second) noexcept { _second = second; }
-				constexpr uint64_t nanoSecond() const noexcept { return _nanoSecond; }
-				void nanoSecond(const uint64_t nanoSecond) noexcept { _nanoSecond = nanoSecond; }
+				constexpr uint32_t nanoSecond() const noexcept { return _nanoSecond; }
+				void nanoSecond(const uint32_t nanoSecond) noexcept { _nanoSecond = nanoSecond; }
 
 				ormTime_t(const char *time) noexcept : ormTime_t{}
 				{
@@ -116,7 +115,7 @@ namespace tmplORM
 					if (time[0] == 0)
 						return;
 					++time;
-					toInt_t<uint64_t> nanoSeconds(time);
+					toInt_t<uint32_t> nanoSeconds(time);
 					if (nanoSeconds.length() <= 9)
 						_nanoSecond = nanoSeconds * power10(9 - nanoSeconds.length());
 					else
@@ -128,7 +127,7 @@ namespace tmplORM
 					fromInt_t<uint16_t, uint16_t> hour{_hour};
 					fromInt_t<uint16_t, uint16_t> minute{_minute};
 					fromInt_t<uint16_t, uint16_t> second{_second};
-					fromInt_t<uint64_t, uint64_t> nanoSecond{_nanoSecond};
+					fromInt_t<uint32_t, uint32_t> nanoSecond{_nanoSecond};
 
 					auto str = makeUnique<char []>(hour.length() + minute.length() +
 						second.length() + nanoSecond.fractionLength(9) + 1);
@@ -203,7 +202,7 @@ namespace tmplORM
 						size_t leapCount;
 					};
 
-					void display() const noexcept { printf("%04u-%02u-%02u %02u:%02u:%02u.%" PRIu64 "\n", _year, _month, _day, _hour, _minute, _second, _nanoSecond); }
+					void display() const noexcept { printf("%04u-%02u-%02u %02u:%02u:%02u.%u\n", _year, _month, _day, _hour, _minute, _second, _nanoSecond); }
 
 					constexpr bool isLeap(const rep_t year) const noexcept
 						{ return (year % 4) == 0 && ((year % 100) != 0 || (year % 400) == 0); }
@@ -255,7 +254,7 @@ namespace tmplORM
 				public:
 					constexpr ormDateTime_t() noexcept : ormDate_t{}, ormTime_t{} { }
 					constexpr ormDateTime_t(const uint16_t year, const uint16_t month, const uint16_t day,
-						const uint16_t hour, const uint16_t minute, const uint16_t second, const uint64_t nanoSecond) noexcept :
+						const uint16_t hour, const uint16_t minute, const uint16_t second, const uint32_t nanoSecond) noexcept :
 						ormDate_t{year, month, day}, ormTime_t{hour, minute, second, nanoSecond} { }
 					ormDateTime_t(const char *dateTime) noexcept : ormDate_t{dateTime}, ormTime_t{dateTime + 11} { }
 
@@ -291,7 +290,7 @@ namespace tmplORM
 						fromInt_t<uint16_t, uint16_t> hour{_hour};
 						fromInt_t<uint16_t, uint16_t> minute{_minute};
 						fromInt_t<uint16_t, uint16_t> second{_second};
-						fromInt_t<uint64_t, uint64_t> nanoSecond{_nanoSecond};
+						fromInt_t<uint32_t, uint32_t> nanoSecond{_nanoSecond};
 
 						auto str = makeUnique<char []>(year.length() + month.length() + day.length() + 1 +
 							hour.length() + minute.length() + second.length() + nanoSecond.fractionLength(9));
